feat(scripts): Add more features like ansi coloring
This commit replaces the bash script with a more powerful Python script
that performs the transformations to the examples READMEs. This new
script comes with the following features added:

- Code blocks ANSI coloring
- Conversion from md H1 to mdx title
- Addition of necessary mdx imports
- Conversion of md quotes to mdx admonitions
- Transform of docs urls to relative syntax

It also adds a guard in the Makefile to only transform into guides the
examples that have been updated, currently marked by the presence of the
substring 'Set metro to Frankfurt, DE'.

Signed-off-by: Alex-Andrei Cioc <andrei.cioc@unikraft.io>